fixed conflict, put code in the correct file and deleted the code from the incorrect file
